Update order_cards.js
Updated to reflect new mandatory field "product_type". This commit reverses the "MAESTRO NFC" field in the "Line on the card" back to "CARD DESCRIPTION" to make it clear that "CARD DESCRIPTION" is what should be replaced by the user to give it a proper name. (Such as MAESTRO NFC or whatever you want to distinguish the card from others!). Also reverts the default PIN back to 1337 to keep that in line with the original version by Crecket.